+++ psl-core/src/main/java/org/linqs/psl/model/rule/WeightedRule.java
@@ -20,7 +20,7 @@ package org.linqs.psl.model.rule;
 public interface WeightedRule extends Rule {
     public boolean isSquared();
 
-    public double getWeight();
+    public float getWeight();
 
-    public void setWeight(double weight);
+    public void setWeight(float weight);
 }
